refactor: Running의 중복되는 로직 제거
