+++ src/main/java/chess/domain/location/LocationDiff.java
@@ -1,7 +1,5 @@
 package chess.domain.location;
 
-import chess.domain.state.Direction;
-
 public class LocationDiff {
 
     private final int diffFile;
