+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,7 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.location.LocationDiff;
-import chess.domain.state.Direction;
+import chess.domain.location.Direction;
 
 public abstract class Piece {
     private final Team team;
