+++ src/main/java/chess/domain/state/Running.java
@@ -1,6 +1,10 @@
 package chess.domain.state;
 
 import chess.domain.Board;
+import chess.domain.location.Direction;
+import chess.domain.location.Location;
+import chess.domain.location.LocationDiff;
+import chess.domain.piece.Piece;
 
 public abstract class Running implements State {
     private final Board board;
@@ -26,5 +30,65 @@ public abstract class Running implements State {
     @Override
     public State end() {
         return new End(getBoard());
+    }
+
+    @Override
+    public State move(Location source, Location target) {
+        Piece sourcePiece = getBoard().getPiece(source);
+        Piece targetPiece = getBoard().getPiece(target);
+
+        checkSourceColor(sourcePiece);
+        LocationDiff locationDiff = source.computeDiff(target);
+
+        checkDirection(sourcePiece, locationDiff.computeDirection());
+        checkDistance(sourcePiece, locationDiff);
+        checkRoute(source, locationDiff);
+        checkTarget(targetPiece);
+
+        if (sourcePiece.isPawn()){
+            checkPawnTargetLocation(targetPiece, locationDiff);
+        }
+
+        getBoard().move(source, target);
+
+        if (targetPiece.isKing()) {
+            return end();
+        }
+        return getOpposingTeam();
+    }
+
+    private void checkPawnTargetLocation(Piece targetPiece, LocationDiff locationDiff) {
+        if (!Direction.isForward(locationDiff.computeDirection()) && !targetPiece.isBlack()) {
+            throw new IllegalArgumentException("[ERROR] 폰은 대각선에 상대 기물이 있을때만 움직일 수 있습니다.");
+        }
+        if (Direction.isForward(locationDiff.computeDirection()) && !targetPiece.isEmpty()) {
+            throw new IllegalArgumentException("[ERROR] 폰은 앞에 기물이 존재하면 직진할 수 없습니다.");
+        }
+    }
+
+    abstract Running getOpposingTeam();
+    abstract void checkSourceColor(Piece piece);
+    abstract void checkTarget(Piece piece);
+
+    private void checkDirection(Piece piece, Direction direction) {
+        if (!piece.isMovableDirection(direction)) {
+            throw new IllegalArgumentException("[ERROR] 해당 위치로 이동할 수 없습니다.");
+        }
+    }
+
+    private void checkDistance(Piece piece, LocationDiff locationDiff) {
+        if (!piece.isMovableDistance(locationDiff)) {
+            throw new IllegalArgumentException("[ERROR] 해당 위치까지 이동할 수 없습니다.");
+        }
+    }
+
+    private void checkRoute(Location source, LocationDiff locationDiff) {
+        Location routeLocation = source.copyOf();
+        for (int i = 0; i < locationDiff.computeDistance() - 1; i++) {
+            routeLocation = routeLocation.add(locationDiff.computeDirection());
+            if (!getBoard().isEmpty(routeLocation)) {
+                throw new IllegalArgumentException("[ERROR] 해당 경로를 지나갈 수 없습니다. ");
+            }
+        }
     }
 }
