+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -3,7 +3,7 @@ package chess.domain.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.location.LocationDiff;
-import chess.domain.state.Direction;
+import chess.domain.location.Direction;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
